requirements.txt
## **Step 3: Create requirements.txt**

```bash
cat > requirements.txt << 'EOF'
openai>=1.0.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
numpy>=1.24.0
pathlib2>=2.3.7
typing-extensions>=4.5.0
EOF